pyproject.toml
[project]
name = "wrappy-crappy"
version = "0.1.0"
description = "Keep your shit together — a security over obscurity wrapper for vibe-coded CLIs"
readme = "README.md"
authors = [
    { name = "Dmitry Aminev", email = "voltterra@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0",
]

[project.scripts]
crpy = "wrappy_crappy.cli:main"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"
